UPDATE project structure frontend entry
Move backend code out of app/ into top-level backend/ and move frontend sources into src/. Update .gitignore with common ignores and add. Add top-level package.json (Electron/Vite build config) and update index.html favicon path. Modify App.jsx to introduce devMode state and pass props to Navbar and Settings.
