--- app/frontend/src/assets/components/Line.jsx
@@ -1,20 +1,0 @@
-import {Line} from 'react-chartjs-2'
-import {Chart as ChartJS, CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend} from "chart.js"
-import { lineChartData } from './Dataset';
-
-ChartJS.register(
-    CategoryScale,
-    LinearScale,
-    PointElement,
-    LineElement,
-    Title,
-    Tooltip,
-    Legend
-);
-
-
-export const LineGraph = () => {
-    const options = {}
-    const data = {lineChartData}
-    return <> <Line options={options} data={data} /></>;
-}
