--- app/frontend/src/assets/components/TopNav.jsx
@@ -1,1 +1,0 @@
-import React from "react";
