--- app/frontend/src/assets/views/Dashboard.module.css
@@ -1,26 +1,0 @@
-@import "tailwindcss";
-/*
-.DashboardElement {
-    background-color:cadetblue;
-    display: flex;
-    justify-content: center;
-    border-radius: 3px;
-    align-items: center;
-    column-span: 1;
-    
-
-
-}
-
-.DashboardLayout {
-    text-align: center;
-}
-
-.Dashboard {
-    min-height: 100vh;
-    width: 100%;
-    align-items: center;
-    justify-content: center;
-    background-color: ;
-}
-    */
