--- app/frontend/src/assets/views/Settings.jsx
@@ -1,11 +1,0 @@
-const Settings = () => {
-  return (
-    <>  
-    <div>
-        <p>Settings</p>
-    </div>
-    </>
-  );
-};
-
-export default Settings;
