+++ backend/sensors/templates/derived_sensor_template.py
