+++ backend/sensors/templates/physical_sensor_template.py
